[springboot-employee-jpa][jenkin] refactor: refactor method deleteCompanyById

+++ src/main/java/com/thoughtworks/springbootemployee/service/CompanyServiceImpl.java
@@ -56,8 +56,12 @@ public class CompanyServiceImpl implements CompanyService {
     }
 
     @Override
-    public void deleteCompanyById(int companyId) {
+    public Company deleteCompanyById(int companyId) {
+        Company company = companyRepository.findById(companyId).get();
+        if (company.equals(null))
+            return null;
         companyRepository.deleteById(companyId);
+        return company;
     }
 
     @Override
